Redirect to hash route for hosted static content
Since we are currently hosting the static built files where
(as far as I know) we can't serve the index.html file for all
requests, we are using hash routing with react-router. This
means that if a user tries to route to a direct, non-hash uri
it shows an error page. This commit adds that error file to
redirect back to a hash uri from index.html so that the react-
router can handle the uri again.