Notify users on all server errors
This is obviously not ideal, however, it is better to
at least notify the user that something has happened
than have no information as to why something is failing.

Alerting here means that all server errors are handled,
and it doesn't reset the promise chain, which was causing
the redux state to be corrupted.